Move PingBadge to uiv3 components
New Compose components belong in uiv3/components per the project
conventions; PingBadge was added to the legacy ui/compose tree. It
still draws on the legacy theme and caption text styles, as the other
uiv3 components do.
