+++ walletkit-chain-zcash/src/main/java/io/horizontalsystems/walletkit/modules/zcashnetwork/ZcashNetworkPage.kt
@@ -48,7 +48,7 @@ import io.horizontalsystems.walletkit.ui.compose.components.HsDivider
 import io.horizontalsystems.walletkit.ui.compose.components.HsIconButton
 import io.horizontalsystems.walletkit.ui.compose.components.HsSwitch
 import io.horizontalsystems.walletkit.ui.compose.components.MenuItem
-import io.horizontalsystems.walletkit.ui.compose.components.PingBadge
+import io.horizontalsystems.walletkit.uiv3.components.PingBadge
 import io.horizontalsystems.walletkit.ui.compose.components.RowUniversal
 import io.horizontalsystems.walletkit.ui.compose.components.VSpacer
 import io.horizontalsystems.walletkit.ui.compose.components.body_jacob
